+++ CHANGELOG.rst
@@ -8,14 +8,15 @@ Notable project changes since release 1.3.1 (2017-02-22).
 Unreleased
 ==========
 
+
 Added
 -----
 
+- DFTB3 coupled-perturbed atom derivatives
 
 
 Changed
 -------
-
 
 
 Fixed